Move BatchedTask to top level
